// archive/sdk-pre-vnext-20260606/src/runtime/generated/runtime/v1/common.ts
export interface ScopedRuntimeBindingAttachment {
    /**
     * @generated from protobuf field: string binding_id = 1
     */
    bindingId: string;
    /**
     * @generated from protobuf field: string binding_handle = 2
     */
    bindingHandle: string;
    /**
     * @generated from protobuf field: string runtime_app_id = 3
     */
    runtimeAppId: string;
    /**
     * @generated from protobuf field: string app_instance_id = 4
     */
    appInstanceId: string;
    /**
     * @generated from protobuf field: string window_id = 5
     */
    windowId: string;
    /**
     * @generated from protobuf field: string avatar_instance_id = 6
     */
    avatarInstanceId: string;
    /**
     * @generated from protobuf field: string agent_id = 7
     */
    agentId: string;
    /**
     * @generated from protobuf field: string conversation_anchor_id = 8
     */
    conversationAnchorId: string;
    /**
     * @generated from protobuf field: string world_id = 9
     */
    worldId: string;
}
// ...
export const ScopedRuntimeBindingAttachment = new ScopedRuntimeBindingAttachment$Type();